sequencing window
Adds a design doc to describe a more flexible sequencing window design
where it can be modified by the chain operator as a `SystemConfig`
config option, similar to other config options. The sequencing
window of remote chains in the derivation pipeline of the local chain
is required to enable the safety of deposits of executing messages in
the context of interop.

While it is possible that interop can ship with a hard assumption on the
value of the sequencing window, we should consider a better way to make
the value legible to the protocol and at the same time make L3s more
stable.