design-docs: update template
Updates the template to include a section about resource usage.
This is important to consider with design docs because blockchain
is a systems engineering discipline and we need to be considerate
of the resource usage of our solutions.

Also move the alternative solutions to below the proposed solution.
This flow makes more sense when reading the document.