interop: Topology and Tx Flow for Interop Chain Operation (#218)
* interop: design doc for topology

Design doc that is meant to define clearly the topology for a cloud
deployment of an OP Stack cluster. This is not meant to define the
only way to deploy a cluster, but is the architecture that we will
do our security review against to ensure that it is not possible to
attack a cluster.

* typo: fix

Co-authored-by: Axel Kingsley <axel.kingsley@gmail.com>

* Add Host Redundancy Topic

* Editorial Pass

* Unsaved edits

* Add Conductor Action Item

* Explicit op-geth mention

---------

Co-authored-by: Axel Kingsley <axel.kingsley@gmail.com>